Continued ch6.
Explored linearity vs nonlinearity in neural networks. Nonlinearity is necessary to make networks capable of taking inputs and extracting 'layers' of information from them. For example, a multi-layer network can take individual pixels as input and determine if the pixels make certain patterns. The presence or absence of these patterns can then be processed to determine if there is a cat in an image.